Added my parts to the game (info on doc)
I added a player, a black floor to make it look like it is ontop of space (later, we can make like a box contraption so it looks more space like). Also, I added the camera to follow the player and for the game to restart when the player collides with an obstacle.